tests: rename test files
