+++ test/testdata/configs/predeclared_custom.yml
