+++ test/testdata/ireturn.go
