+++ test/testdata/nonamedreturns_custom.go
@@ -1,5 +1,5 @@
 //golangcitest:args -Enonamedreturns
-//golangcitest:config_path testdata/configs/nonamedreturns.yml
+//golangcitest:config_path testdata/configs/nonamedreturns_custom.yml
 package testdata
 
 import "fmt"
