+++ test/testdata/testifylint_bool_compare_only.go
